New series with minimal points (mostly octagonal layout) from Reconstruct.
These new series were created in reconstruct with simplify on and off.
The tracing in both cases was done using the Closed Point-by-Point Drawing
tool on the vertex points (or inscribed octagon) of the shapes. When the
simplify option ("Automatically simplify new closed traces ...") is
enabled, Reconstruct.exe seems to insert extra points. The locations of
the inserted points seems somewhat arbitrary as shown in the images.